fix(sr): 修复预加载空操作等问题并重构代码
### 问题修复
- **SRPreloadDispatcher**: 修复预加载协程为空操作的 bug，添加 `onPreloadRequested` 回调机制，
  并在 ReaderViewModel 翻页时触发预加载（`onPageChanged` 之前从未被调用）
- **RealESRGANProcessor**: 移除弃用的 `finalize()` 方法，资源清理已由显式 `release()` 路径保证
- **SuperResolutionManager.process()**: 添加 `withTimeout(120s)` 超时保护，防止 SR 推理无限阻塞

### 内存与并发安全
- **realesrgan_wrapper.cpp**: `weight_sum` 由 `uint16_t` 改为 `uint32_t`，消除超大图片 tiling 溢出风险

### 代码质量
- **manga_bw_postprocessor.cpp**: 提取 `applyThreshold()` 消除 `binarizeEnhance` 两分支的代码重复
- **SrStatusIndicator.kt**: 将 `labelText` 延迟计算移到 `ICON_AND_TEXT` 分支内，避免 `ICON_ONLY` 模式不必要开销
- **SRDiskCache.kt**: 优化缓存逐出策略，添加 TODO 异步淘汰注释
- **ModelManager.kt**: 缓存 `modelsDir` 实例，避免重复构造
- **.gitignore**: 清理构建调试文件模式（`*.txt`, `*.ps1`, `ci_logs/`）

+++ core/superresolution/src/main/cpp/realesrgan_wrapper.cpp
@@ -143,7 +143,7 @@ bool RealESRGANWrapper::process(const ncnn::Mat& inimage, ncnn::Mat& outimage) {
         memset(ch_data, 0, out_w * out_h * sizeof(float));
     }
 
-    std::vector<uint16_t> weight_sum(out_w * out_h, 0);
+    std::vector<uint32_t> weight_sum(out_w * out_h, 0);
 
     int tile_size = tilesize > 0 ? tilesize : 200;
     int min_overlap = 12;
@@ -313,7 +313,7 @@ bool RealESRGANWrapper::process(const ncnn::Mat& inimage, ncnn::Mat& outimage) {
 
                         dst_row[dx] += src_row[sx] * wt;
                         if (c == 0) {
-                            weight_sum[out_idx] += static_cast<uint16_t>(wt * 256 + 0.5f);
+                            weight_sum[out_idx] += static_cast<uint32_t>(wt * 256 + 0.5f);
                         }
                     }
                 }
@@ -325,7 +325,7 @@ bool RealESRGANWrapper::process(const ncnn::Mat& inimage, ncnn::Mat& outimage) {
         for (int y = 0; y < out_h; y++) {
             float* dst_row = static_cast<float*>(outimage.channel(c).row(y));
             for (int x = 0; x < out_w; x++) {
-                uint16_t ws = weight_sum[y * out_w + x];
+                uint32_t ws = weight_sum[y * out_w + x];
                 if (ws > 0) {
                     dst_row[x] = dst_row[x] * 256.0f / static_cast<float>(ws);
                 }
